Initial commit
This will be a basic skeleton chef repository for new projects. It supports
cookbooks and roles as normal. Node configuration (run lists, per node
attributes) is in nodes/nodename.json.

The run_chef.sh is a simple script intended to do a single chef run (but there
are options in place for running as a daemon), and it picks the node name just
by running 'hostname'. We may need to improve on this later.